requirements.txt
alembic==1.18.4
bcrypt==3.2.2
cryptography==46.0.5
fastapi==0.129.0
idna==3.11
Jinja2==3.1.6
psycopg==3.3.3
pydantic==2.12.5
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
SQLAlchemy==2.0.46
uvicorn==0.40.0
